[Front] refactor(skills): rename remaining skill configuration routes from `assistant` to `skills`
