+++ front/pages/api/w/[wId]/skills/[sId]/editors.ts
