+++ front/pages/api/w/[wId]/skills/[sId]/index.test.ts
